requirements.txt
numpy==1.21.2
opencv-python
pytorch==1.10.0
torchvision==0.11.1
umap-learn
matplotlib==3.4.3
pandas==1.3.4
ipywidgets>=7.5
tqdm
wandb
fiftyone
requests
libwebp

# $ conda install -c conda-forge accimage
# or refer to
# https://gist.github.com/brhcriteo/f659c2eecd7f4a45b12a6988559865a7